Update run_1stLevel_Analysis.sh
I'm following the owner of this code (andrewjahn)'s fMRI Tutorial but there was a problem with my subj-08 folder in that the bash make_FSL_Timings.sh command did not create timing files for my subj-08. So I decided to start over with subj-01, which the bash command did create timing files for. That's why I wanted to edit this file to fit subj-01 and the fractional intensity that I found worked the best for me. Please let me know if there is an error. Thank you!